feat: move menu icon to header
