+++ core/gatsby-theme-docz/src/components/Sidebar/styles.js
@@ -9,7 +9,7 @@ export const global = {
 export const overlay = ({ open }) => ({
   zIndex: 999,
   position: 'fixed',
-  top: 104,
+  top: 88,
   right: 0,
   bottom: 0,
   left: 0,
@@ -37,7 +37,7 @@ export const wrapper = ({ open }) => ({
     zIndex: 9999,
     display: 'block',
     position: 'fixed',
-    top: 104,
+    top: 88,
     left: 0,
     bottom: 0,
     width: 256,
